ch05/ex05/main.go
package main

import (
	"bufio"
	"net/http"
	"strings"

	"golang.org/x/net/html"
)

func CountWordsAndImages(url string) (words, images int, err error) {
	resp, err := http.Get(url)
	if err != nil {
		return
	}
	doc, err := html.Parse(resp.Body)
	if err != nil {
		return
	}
	words, images = countWordsAndImages(doc)
	return
}

func countWordsAndImages(n *html.Node) (words, images int) {
	if n == nil {
		return
	}
	if n.Type == html.ElementNode && n.Data == "img" {
		images++
	}

	if n.Type == html.TextNode {
		r := bufio.NewScanner(strings.NewReader(n.Data))
		r.Split(bufio.ScanWords)
		for r.Scan() {
			words++
		}
	}

	w, i := countWordsAndImages(n.FirstChild)
	words += w
	images += i

	w, i = countWordsAndImages(n.NextSibling)
	words += w
	images += i

	return
}
